requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
et_xmlfile==1.1.0
fastapi==0.116.1
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
h3==4.3.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.4.2
lightgbm==4.3.0
numpy==2.2.0
openpyxl==3.1.2
pandas==2.2.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.2
six==1.16.0
sniffio==1.3.1
starlette==0.47.2
threadpoolctl==3.5.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
uvicorn==0.35.0
